refactor: take socket event handlers to separate files (#255)
* refactor: place sockets in seperate folder and remove console messages

* fix: unreachable files

* refactor: return err rather than using console.log

* refactor: return err rather than using console.log
